feat: add checklist button and progress tooltip
refactor: squad steps logic to hook to be able to reuse
